streamlit_app/requirements.txt
streamlit>=1.30
matplotlib
numpy
rasterio
Pillow
